backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Messias Junior <juniormj1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
junos-eznc = "^2.6.3"
requests = "^2.27.1"
SQLAlchemy = "1.4.7"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
asyncpg = "^0.26.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytz = "^2022.2.1"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.10.2"}
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
pylint = "^2.12.2"

[tool.poetry.scripts]
backend = "backend.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
